+++ services/core-service/src/apps/apps.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common"
 import { AssetManagerModule } from "./assetmanager/assetmanager.module"
 import { DebtTrackModule } from "./debttrack/debttrack.module"
-import { GoalModule } from "./goal/goal.module"
+import { GoalManagerModule } from "./goalmanager/goalmanager.module"
 import { DiscoverModule } from "./discover/discover.module"
 import { ExpenseTrackModule } from "./expensetrack/expensetrack.module"
 import { TaxAdvisorModule } from "./taxadvisor/taxadvisor.module"
@@ -13,7 +13,7 @@ import { PlannerModule } from "./planner/planner.module"
     AssetManagerModule,
     DebtTrackModule,
     ExpenseTrackModule,
-    GoalModule,
+    GoalManagerModule,
     DiscoverModule,
     TaxAdvisorModule,
     CashFlowModule,