+++ services/core-service/src/auth/dto/google-oauth.dto.ts
@@ -2,5 +2,5 @@ import { IsNotEmpty } from "class-validator"
 
 export class GoogleOAuthDto {
   @IsNotEmpty()
-  readonly token: string
+  readonly code: string
 }