+++ services/core-service/src/shared/llm/llm.service.ts
@@ -8,13 +8,13 @@ export class LLMService {
 
   getLLM() {
     return new ChatOpenAI({
-      model: config.GPT_OSS_MODEL_NAME,
+      model: config.AI_MODEL_MODEL_NAME,
       temperature: 0.8,
       topP: 0.8,
-      apiKey: config.GPT_OSS_API_KEY,
+      apiKey: config.AI_MODEL_API_KEY,
       configuration: {
-        baseURL: config.GPT_OSS_DEPLOYMENT_URI,
-        apiKey: config.GPT_OSS_API_KEY,
+        baseURL: config.AI_MODEL_DEPLOYMENT_URI,
+        apiKey: config.AI_MODEL_API_KEY,
       },
     })
   }