+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
+## Version 3.6.1
+
+- Implemented Google OAuth with PKCE flow and deprecated legacy flow
+- Migrated from GPT OSS to GPT 4o for better reasoning
+- Added intelligence layer to all new apps
+
 ## Version 3.6.0
 
 - Code refactoring & optimization
-- Major changes in UI & core service
+- Changes in date handling in schemas across services
+- UI change for Month navigation in Planner & ExpenseTrack
+- Rebranding of 2 applications
+- Removed legacy OTP based authentication
 
 ## Version 3.5.0